+++ packages/client-ledger/src/lib/config.ts
@@ -59,7 +59,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.0.0",
-    sdkVersion: "1.0.0-alpha.16",
-    genVersion: "2.385.2",
-    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.16 2.385.2 0.0.0 @formance/sdk-ledger",
+    sdkVersion: "1.0.0-alpha.17",
+    genVersion: "2.387.0",
+    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.17 2.387.0 0.0.0 @formance/sdk-ledger",
 } as const;